finding buffers-1 (P1): restore int64 idx_mapping on the MRV2 pre-allocated buffer
vendor #35561 (dc576b0450) introduced
`self.idx_mapping = CpuGpuBuffer(max_num_reqs, dtype=torch.int32, ...)`.
That PR is still OPEN upstream and was authored against the pre-#51210 tree,
where request indices were int32 on both sides:

    #35561: idx_mapping_np = np.fromiter(..., dtype=np.int32)
            self.idx_mapping.np[:num_reqs] = idx_mapping_np

Merged #51210 widened them (np.int32 -> np.intp, torch.int32 -> torch.int64).
Our tree carries #51210: `prepare_batch_state` builds idx_mapping_np as
np.intp and `InputBatch.make_dummy` builds `torch.arange(num_reqs,
dtype=torch.int64)`. The vendored buffer therefore re-narrowed only the
serving-side tensor: dummy/profile/capture batches carried int64 request
indices while every real step carried int32 for the same argument, and the
staging write silently downcast intp -> int32 on every step.

Fix: allocate the idx_mapping buffer as int64. expanded_idx_mapping follows
automatically (`idx_mapping.new_empty` in expand_idx_mapping).

Other five buffers checked against the allocations they replaced — all
correct, no change:
  cu_num_logits  <- np.empty(num_reqs+1, np.int32) + async_copy_to_gpu   int32
  arange_np      <- np.arange(num_reqs+1, dtype=np.int32)                int32
  arange_gpu     <- torch.arange(num_reqs+1, dtype=torch.int32)          int32
  zeros_gpu      <- torch.zeros(num_reqs, dtype=torch.int32)             int32
  query_start_loc<- aliases input_buffers.query_start_loc (int32),
                    staged from np.empty(max_num_reqs+1, np.int32)       int32

Same-pass consequence: spec_decode_rejection_warmup built its dummy
idx_mapping / expanded_idx_mapping as int32, which after this change would
compile a Triton specialization no real batch uses (warmup silently useless).
Widened to int64 with the reason named.

+++ vllm/model_executor/warmup/spec_decode_rejection_warmup.py
@@ -140,9 +140,12 @@ def spec_decode_rejection_warmup(worker: Worker) -> None:
                 # positions buffer is int64; idx_mapping / expanded_idx_mapping /
                 # expanded_local_pos are the runner's int32 buffers.
                 pos=torch.zeros(num_logits, dtype=torch.int64, device=device),
-                idx_mapping=torch.zeros(num_reqs, dtype=torch.int32, device=device),
+                # int64 to match the serving dtype of idx_mapping /
+                # expanded_idx_mapping (#51210); an int32 warmup would compile
+                # a Triton specialization the real batches never use.
+                idx_mapping=torch.zeros(num_reqs, dtype=torch.int64, device=device),
                 expanded_idx_mapping=torch.zeros(
-                    num_logits, dtype=torch.int32, device=device
+                    num_logits, dtype=torch.int64, device=device
                 ),
                 expanded_local_pos=torch.arange(
                     num_logits, dtype=torch.int32, device=device
